[feat:(refactor)]: Complete refactor android folder
