+++ example/android/app/build.gradle
@@ -68,8 +68,8 @@ def jscFlavor = 'org.webkit:android-jsc:+'
 
 android {
     ndkVersion rootProject.ext.ndkVersion
-    compileSdkVersion rootProject.ext.compileSdkVersion
-    namespace "com.rnmaplibreexample"
+    compileSdk rootProject.ext.compileSdkVersion
+    namespace "com.rnvietmapexample"
 
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
@@ -77,7 +77,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.rnmaplibreexample"
+        applicationId "com.rnvietmapexample"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
