+++ example/android/app/src/release/java/com/rnvietmapexample/ReactNativeFlipper.java
@@ -4,7 +4,7 @@
  * <p>This source code is licensed under the MIT license found in the LICENSE file in the root
  * directory of this source tree.
  */
-package com.rnmaplibreexample;
+package com.rnvietmapexample;
 
 import android.content.Context;
 import com.facebook.react.ReactInstanceManager;
