+++ example/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "RNMaplibreExample",
-  "displayName": "RNMapLibreExample"
+  "name": "RNVietmapExample",
+  "displayName": "RNVietmapExample"
 }
